// src/utils/types/Types.ts
// Components variants types

export type ElevationVariant = 'card' | 'modal' | 'notification' | 'box' | 'warning' | 'panel'

export type ButtonVariant = 'text' | 'outlined' | 'contained' | 'neutral'

export type TypographVariant = 'headline' | 'subtitle' | 'body' | 'caption'

export type TypographColorsOptions =
    | 'primary'
    | 'secondary'
    | 'success'
    | 'error'
    | 'warning'
    | 'neutral1'
    | 'neutral2'
    | 'default';